requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow
torch
transformers
opencv-python-headless==4.8.1.78
numpy
sentencepiece
diff-match-patch
timm